+++ src/asmout.c
@@ -2,6 +2,27 @@
 
 #include "asmout.h"
 
+/*
+ * Instruction conversion
+ * */
+
+/*
+ * Run
+ * */
+
+void convertToAsm(AsmOut *out) {
+	
+}
+
+
+/*
+ * Initialization & exiting
+ * */
+
+int getRoughFileSize(AsmOut *out) {
+	return out->parser->bufferSize*2;
+}
+
 void freeAsm(AsmOut *out) {
 	fclose(out->asmOut);
 	free(out->functions);
@@ -12,9 +33,25 @@ void initAsmOut(struct parserData *parser, AsmOut *output) {
 	int i;
 	output->functions = (Function *)malloc(sizeof(Function)*parser->totalFunctions);
 
-	char fbuf[100];
-	snprintf(fbuf, sizeof(fbuf), "%s.s", parser->fData->fileName);
-	output->asmOut = fopen(fbuf, "wr");
+	char *fileName = strtok(parser->fData->fileName, ".");
+	strcat(fileName, ".s");
+
+	output->asmOut = fopen(fileName, "r");
+	/*File already exists*/
+	if(output->asmOut!=NULL) {
+		char rm[100];
+		snprintf(rm, sizeof(rm), "rm %s", fileName);
+		system(rm);
+	} 
+	fclose(output->asmOut);
+	
+	/*reopen with write mode/create file*/
+	output->asmOut = fopen(fileName, "wr");
+
+	int roughEstimate = getRoughFileSize(output);
+	output->parser->output.asmOutBuffer = (char *)malloc(sizeof(char)*roughEstimate);
 
 	/*Don't forget output buffer is in parser->output.asmOutBuffer*/
 }
+
+