+++ src/cpu.h
@@ -1,0 +1,5 @@
+/*Copyright (c) 2024 Tristan Wellman*/
+#ifndef CPU_H
+#define CPU_H
+
+#endif