+++ src/werror.h
@@ -10,6 +10,8 @@ struct wErrors errors;
 
 #define WERROR_MISSING_BRACKET 1
 #define WERROR_MAIN 2
+#define WERROR_UNDEFINED_INSTRUCTION 3
+#define WERROR_UNINITIALIZED_VARIABLE 4
 
 #define WLOG_WERROR(werror, file, line, function, ...) 		\
 	fprintf(stderr, "%s \033[31m"#werror"\033[0m in %s:%d - %s\n%s",		\