+++ changelog.md
@@ -1,12 +1,20 @@
-# 11/23/24 - Total rewrite
+# Alpha v0.0.1 - 2024
+
+## 11/23/24 - Total rewrite
 
 * This "change" is deleting all of what was and adding this initial commit
 * functional parser.
 
-# 11/25/24 - Design, parser, asmOut
+## 11/25/24 - Design, parser, asmOut
 
 * Started the asm output file.
 * finished instruction parsing, this pretty much completes the parser.
 * some design additions, added to the design file documentation.
 * asm "example" test coorilating to the helloworld.well program.
+
+## 11/26/24 - parser, asmOut
+
+* Fixed some asmOut initialization issues.
+* Did variable parsing.
+* added some Werrors, fixed tab eater.